Move js/utils.ts to utils/common.ts
