+++ src/js/App/GlobalFilter/constants.ts
@@ -2,8 +2,8 @@ import omit from 'lodash/omit';
 import flatMap from 'lodash/flatMap';
 import memoize from 'lodash/memoize';
 import { AAP_KEY, MSSQL_KEY, SID_KEY } from '../../redux/globalFilterReducers';
-import { getUrl } from '../../utils';
 import type { Group, GroupFilterItem } from '@redhat-cloud-services/frontend-components/ConditionalFilter';
+import { getUrl } from '../../../utils/common';
 
 export const INVENTORY_API_BASE = '/api/inventory/v1';
 export const workloads = [
