+++ src/js/App/Header/Tools.js
@@ -9,7 +9,7 @@ import ToolbarToggle from './ToolbarToggle';
 import HeaderAlert from './HeaderAlert';
 import { useSelector } from 'react-redux';
 import cookie from 'js-cookie';
-import { getSection, getUrl, isBeta } from '../../utils';
+import { getSection, getUrl, isBeta } from '../../../utils/common';
 import { spinUpStore } from '../../redux-config';
 import classnames from 'classnames';
 import { useIntl } from 'react-intl';
