+++ src/js/App/RootApp/RootApp.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Router } from 'react-router-dom';
 import { HelpTopicContainer, QuickStartContainer } from '@patternfly/quickstarts';
 
-import { isBeta } from '../../utils';
+import { isBeta } from '../../../utils/common';
 import chromeHistory from '../../../utils/chromeHistory';
 import { FeatureFlagsProvider } from '../FeatureFlags';
 import IDPChecker from './IDPChecker';
