+++ src/js/App/RootApp/ScalprumRoot.js
@@ -10,7 +10,7 @@ import NavLoader from '../Sidenav/Navigation/Loader';
 import { usePendoFeedback } from '../Feedback';
 import { toggleFeedbackModal } from '../../redux/actions';
 import historyListener from '../../../utils/historyListener';
-import { isFedRamp } from '../../utils';
+import { isFedRamp } from '../../../utils/common';
 import { SegmentContext } from '../analytics/segment-analytics';
 import LoadingFallback from '../../../utils/loading-fallback';
 import { clearAnsibleTrialFlag, isAnsibleTrialFlagActive, setAnsibleTrialFlag } from '../../../utils/isAnsibleTrialFlagActive';
