+++ src/js/App/Routes/QuickstartsCatalogRoute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { getUrl } from '../../utils';
+import { getUrl } from '../../../utils/common';
 import QuickStartCatalog from '../QuickStart/QuickStartCatalog';
 import { useIntl } from 'react-intl';
 import messages from '../../Messages';
