+++ src/js/App/analytics/segment-analytics.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useEffect, useRef } from 'react';
 import { AnalyticsBrowser } from '@segment/analytics-next';
-import { getUrl, isBeta, isProd } from '../../utils';
+import { getUrl, isBeta, isProd } from '../../../utils/common';
 import { useSelector } from 'react-redux';
 import { ChromeUser } from '@redhat-cloud-services/types';
 import { useLocation } from 'react-router-dom';
