+++ src/js/jwt/insights/entitlements.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { ServicesApi } from '@redhat-cloud-services/entitlements-client';
-import { deleteLocalStorageItems, lastActive } from '../../utils';
+import { deleteLocalStorageItems, lastActive } from '../../../utils/common';
 import { bootstrapCache } from '../../../utils/cache';
 
 const BASE_PATH = '/api/entitlements/v1';
